requirements.txt
Django>=2.0; python_version>='3.6'
djangorestframework>=2.4.3
django-redis>=4.5.0
model_mommy>=1.4.0